wings_we:rebuild/1: Optimize vertex position GC
The vertex position GC part of the rebuild operation will be slow
if there are many more vertex positions in the 'vp' table than
entries in the 'vc' table (for example, when separating an object
with many sub-objects) because the 'vp' table is converted to
a list which must be traversed to discard unused positions.

Fix that by looking up each vertex in the 'vc' table in the original
'vp' array. That will be slightly slower if the tables have the
same number of entries, but much faster if there are many more
entries in the 'vp' table.

NOTE: The Separate command is now significantly faster when run
on objects with many sub-objects. (Thanks to ycarry for noticing
the slowness.) [bjorng]