Enable FreeBSD Compile
Added include and library paths for opengl and jpeg.

Changed make_installer to handle FreeBSD's libSDL.so, which is not a
symbolic link. Also switched to 'bash' because the built-in pwd command
in 'sh' produces a minor error if the wings directory is a symbolic
link.

NOTE: Wings now compiles cleanly on FreeBSD.