Validate hole faces when loading wings files
While at it, rearrange and comment the code in share_list_2 to
make it clearer.

NOTE: Wings will now validate holes when loading .wings files,
and remove from the list of holes any non-existent face or
face that is not invisible. (Importing objects with holes and
saving in a previous release of Wings could cause such
inconsistencies, as described in the previous release note.)
[bjorng]