Fix crash in wings_we:merge/{1,2}
wings_we:merge/{1,2} can sometimes crash because of duplicate keys
in the vertex or edge table. The reason is that merge_renumber/3
can fail to realize that two objects has overlapping ranges if
merge_bounds/2 returns negative lower bounds (causing the list
of Wes to be sorted in the wrong order).

Fix this problem by returning a lower bound of 0 for objects
that have invsible faces.