Teach Wings3D.exe to pass on the AppData location
Using the registry on Windows to find out the AppData location
(i.e. the location where application can store their own data,
typically c:/Documents and Settings/<user>/Application Data)
is not supported, and Windows Vista and Windows 7 store that
information in a different place in the registry than Windows XP.

Therefore, use the SHGetFolderPath() function. It is deprecated,
but the recommended SHGetKnownFolderPath() function is only
available in Windows Vista and later. Here is the documentation:

  http://msdn.microsoft.com/en-us/library/bb762181(VS.85).aspx

Pass on the AppData to Wings using

     -extra "<AppDataLocation>"

last on the command line (it can be retrieved using
init:get_plain_arguments/0).