Created a 3d file format converter script.
A script can be used from the command line, can be used to convert
models from other scripts or makefiles or for testing wings import export
code.

On windows invoke with ERLPATH/bin/escript wings_convert ...

The script needs more testing, I don't think textures work yet and
lights, and stuff but it is a start.