MacOS X: Put the preference file in a sub-directory
The preference files for Wings3d on MacOS X used to be created
directly in the "Library/Preferences directory". Now that is fine
as long as there can only be one preference file, but since we
soon will allow the user to have multiple named preference
files, that could cause preferences files for other applications
to be overwritten. Therefore we should put the preference file
in the directory "Library/Preferences/Wings3D".