Unix: Put the preference file in a sub-directory
The preference file for Wings3D on Unix systems (inluding Linux) used
to be named "$HOME/.wings". Now that is fine as long as there can only
be one preference file, but since we soon will allow the user to have
multiple named preference files, they would all be stored directly
in the home directory, which could turn out to be messy. Therefore
we should put the preference file in "$HOME/.wings3d/preferences".