Tried to fix windows unc path issue
NOTE: Fix unc paths on windows
(Thanks to jentzenm.unofficialsony that reported this bug) [Micheus]

When trying to saving a wings scene file to a path on the network using unc path, an error occurs (Save failed: no such file or directory) and the file was not saved. For some reason the first slash is removed from the path name when Save dialog returns the file name in the function wings_file:export_filename_1/2(after executed wings_plugin:call_ui()). The same occurs with Open dialog too. It seems to be a Erlang issue (I'm not sure). I did a workaround, then I could Save/Export/Import using an "unc" path without any problem. (Thanks to jentzenm.unofficialsony that reported this bug) [Micheus]