Add new by / faces marked  for by Menu
Ability to select face by any face colored,
faces of a particular color, or uncolored faces
(which is really just a inverse of first option).

For users wanting more than just hard-edges or
charts to delineate areas of interest ... colors
become another simple way of marking faces and boundaries
with this addition.