Refactor pointer API as the first step towards VBO support
Currently, geometry is drawn using vertex buffer arrays captured
into a display list. There are several problems doing it like that,
both in the short term and in the long term.

In the short term, displays lists are not very efficiently handled
in some OpenGL drivers. In the long term, display lists are deprecated
in Open GL 3.x and later, and not supported at all in OpenGL 3.x
on the Mac.

Therefore, we will need to move away from using display lists and
start using vertex buffer objects (VBOs).

The current #vab{} record that supports vertex buffer arrays can
be refactored to support VBOs instead. VBOs are similar to
vertex buffer arrays, but are OpenGL objects that needs to be
garbage collected.

As the first step in the refactoring, we will need to hide more
details of the #vab{} record. The first step (this commit) is
to replace all the functions that enable or disable pointers
by two functions:

   enable_pointers(#vab{}, [Extra]).
   disable_pointers(#vab{}, [Extra]).

By using one call to enable all pointers, we can later make sure that
the buffers are set up in the correct order (the VBO must be bound
before any pointers start to reference it). Since the calls are passed
the entire #vab{} record, we can later add more fields to handle VBOs
into the #vab{} record.