Fixed some bugs reported for Wings3d 1.5.2
During some tests with Collada plugin - because of some issues related to the normals exported - it was noticed some inconsistences in the file generated. I used the smllint parser with Collada 1.4.1 schema (on-line | http://collada.org/mediakiki/index.php/Schema_validation):
- the "library_image" session has been exported with no images/texture in the scene and generating an inconsistent id (ID="-img") for the "image" session;
- for the material ("phong"), the diffuse attribute has been written in the wrong order.
- the parser has pointed that the "node"'s id was not unique. It was noticed that it was the same used for its "geometry"'s id.

In order to applay the fixes it was used the "COLLADA Digital Asset Schema Release 1.4.1 Specification (2nd Edition)" and parsed on-line, as well the ".dae" file was read using a 3rd software.

NOTE: Fixed an issue related to edges highlight in Tweak and Sculpt that was crashing wings when the object have hidden faces or holes. Thanks to Extrudeface for report it.

NOTE: Fixed a issue related to Select Group and Saved Camera windows not be drawn using the alpha settings as defined in the preferences. Thanks to Justanother1 for report it.

NOTE: Fixed the duplicated IDs in the language file for "Rename" operations available in the "Geometry Graph" window. Thanks to TulipVorlax.

NOTE: Fixed some issues related to images of materials exported by POV-Ray plugin. Thanks to RyMopar that reported it.

NOTE: Fixed an error in the "wings_u:relative_path_name" routine that was rebuilding the full path in a wrong way.