It was added the possibility of define all the OpenGL settings for default material. In the Material Editor dialog the user can change those fields and use a new button Save for save in the preferences the new material properties. Other button - Restore - enable the user to reset to Wings3d default material properties that is hard coded. These new informations are mantained by a new preferences variable preserving the compatibility previous preference file.
Also, it was changed a little the way how Wings3d read/import the default material from an existent file. If the only difference between both default materials is the OpenGL parameter, then the material being read is replaced by the current one, otherwise it create a new variation of that material name as already has been done. Thanks to MikeJ and oort for the suggestions and tests. [Micheus]