Gently quit, fix hotkey/events
If erlang quits (on mac at least since it runs "main" thread)
before the ports is closed, erlang exit and wx cleanup
both deletes memory from two threads which causes double deletes
and crashes when closing wings.

First close the window, so the driver do not need do so much
cleaning before exiting, since the delete/destory is
async do init:stop() and wait for threads to exit.

Fix hotkey descriptions on Mac and Windows.

Fix char events for modifers and menu shortcut displays.

Fix it by translating from menu events to keyevents,
so we can handle the keyevents instead, since this cannot
be turned off in wxWidgets.

Windows hotkeys must be quoted, otherwise we do not get events
when mouse is grabbed for keys that are accelerators.

Mac modifier command as meta in menues.

sdl mapped command as meta and ctrl as ctrl,
but wxWidgets maps command as ctrl and ctrl as rawctrl.

There might be more bugs in this area...

Delay menubar frame attach until after the menus are created
so that about and preferences goto the correct place.

Request antialising on wxWidgets-3.0