Replaced the default Windows' Open dialog for Font selection.
Due to the impossibility to see the .ttf font file in the Windows7's Open dialog box and the fact of the selected file in the WindowsXP's Open dialog box isn't updated in the Text primitive dialog box, I did these changes in order to enable the old style Font dialog box that enable us to see and select the font correctly. This dialog will be enabled for any Windows OS version.

NOTE: Fixed an issue related to font selection under Windows 7 - fonts' name aren't displayed.