Made the grid size customizable
Many users have asked for an option for customize the grid size. So, now users can define the grid size in the Preferences dialog (General->Grid).
The camera distance is set to 50% of the grid size with a minimum value of 8 wings3d units.

NOTE: Added option to customize the grid size.