Fix the 'make macosx' and 'make unix' targets
Packaging a release is messy with different shell scripts for different
platforms. Unify the packaging script into an escript called tools/release.

The 'makeself' script needed to pack the Unix release is quite small, so
we can as well put a copy of it in the 'unix' sub-directory so that we
don't need to download and install it ourselves.