Add ww_color_ widgets
To be able to select which color chooser to use and support palette
initialization. This is not supported by wxWidgets variant, but
needed if user changes the palette.

Add a color slider and use it.

Tab traversal and keyinput in color_slider
doesn't work on windows and wxWidgets-2.8

Add custom_gl to dialogs and make wx material window