New Absolute Command - Rotate
A new resource Absolute Command->Rotate was added. It intend to give us precision in rotation operations we need to rotate elements to be aligned with a vector or plane. There are three option LMB,MMB and RMB, which MMB works like LMB plus run a snap operation and RMB - similar to Snap/Move - show up a dialog which will provide array options. Thanks to Arg Arg for help testing it [Micheus]

The wings_vec.erl was changed to enable the user change the "direction" of an "orientation" vector by adding the option "[2] Invert" to the text in the right side in the information line. The direction is something important for this new command. It was also applied the fix for the getting axis from three vertices of two or three objects.