Fixed normalization of vectore in move/scale/rotate
Fixed a bug with the routine to get vertex and point for
move/scale/rotate operations. When the vector was obtained from vertex
on two or three objects we could get undesired result during the these
operations.

NOTE: Fixed a bug with the routine to get vertex and point for
move/scale/rotate operations. (Thanks PuzzledPaul)