Let wings calculate modelview and perspective view matrices
Prepare for the future OpenGL versions and a try at an optimization
by avoiding a thread communication when reading the matrices from opengl.

Change order in matrix operations and opengl equiv ortho and perspectiv

Change the mul order for rotate, translate and scale to let it behave as
opengl does. The mul operation was changed in previous commit.

This breaks backwards compatibility.

Also add ortho and perspectiv with the same arguments as the gl/glu equivalents.

Calculate pick matrices without opengl