POV-Ray and Kerkythea plugins ported to use new wx module
Finished porting POV-Ray and Kerkythea plugins to use the new UI - wx.

It was needed to add some changes to wings_dialog.erl in order to enable us to
hide/show controls. To allows us that without create a new control and preserve
the same dialogs definitios already in the source files it was added a key
management to v/hframe control. If a key is set in the flags then we use a
wxPanel with this key as its ID.

The Texture tab in POV-Ray uses that because we have many differents group of
parameters for each different Type. In order to keep the dialog height short we
need to leave only one of these group visible - the greater one - that is why
it was needed to add new flag option for v/hframe: {show,true/false}.

I also added support to .lang file syntax in Kerkythea plugin.