Unicode support to the preferences and language files.
The current code to load preferences file by using "file:consultant" is breaking
Wings3d when special characters are found inside on it.

It was reviewed the previous implementation as requested by dgud in order to use
a short code to fix the unicode problem.
The preference file - before written in latin1 - will be converted and rewritten
using the UTF8 encoding.

OBS:
1) The changes to tool.erl was needed because of the similar problem when reading
some *.lang files (like portuguese or *_tw) and was not possible to generate the
installer.

2) If the new encoded file CONTAINS unicode chars it will not be loadable by
any previous version of Wings3D (no crash), otherwise the file will still be
"valide". That was what I noticed during my tests.