POV-Ray and Kerkythea plugins final checking after ported to use wx module
- POV-Ray and Kerkythea plugins: it was tried to make the two code more
similar by using same functions structure and name always as possible. That
was done in order to try make and easy read the code and the maintence process;
- Yafray plugin removed from makefile and plugins_src folder. The old yafray
was removed due it has became obsolete. It will be replaced by the Yafaray
plugin in development by oort;
- wings_dialog.erl has Show/3 function changed to not call wxWindow update for
every control we show. It was added an Update/3 function that allow the coder
to call wxWindow updates only after all window has been shown/hidden by code;
- wings_u.erl had the relative_path_name routine updated in order to fix a bug
found during the plugins tests.