The table control now returns a selection list and the input values
The table control also returns a tuple value composed by a list of selected
item index and the original data values. As it used to be in the ESDL version.

The wings_dialog.erl was changed in order to enable the user to set the
max number of row in the table for a best look-and-feel. The best way to
calculate it would be getting the font size used by the control or the
its row height, but I couldn't find a direct/safe way get it.

We still need to find a way to change the values during a control event - as
it could be done in the past with {update,Store}. Dynamics dialogs requires
that all the time.