Fixed the popup menu location
When the popup menu was invoked close to 3rd part donw in the W3D client area
it could be displayed Down from that position and some options became hidden
depending on where the W3D windows is on the windows desktop. The fix check
for the desktop height before set the menu position.