wpc_ambocc: Properly delete #vab{} records
For the moment, wings_draw_setup:delete_vab/1 is a dummy, but it will not
be in a future commit. Refactor the Ambient Occlusion plug-in to properly
delete #vab{} records when they are no longer used. Even though the call
to wings_draw_setup:delete_vab/1 is a dummy, it will actually ensure that
the #vab{} record is not garbage collected away too early.

Also remove the incorrect usage of wings:keep/1. Note that the list
passed to wings:keep/1 will only be a list of 'ok' atoms.