New Merge dialog
- It was added a dialog to the Merge option in which the user can choose what
he/she wants to import.

In order to allow the dialog controls act dynamically in accord with the
user choices it wass added code to change the checkbox and radiobox values
on the fly.
A new flag was also added to the Table control allow us to set the selection
mode to single selection. Not used in the current code, but it can be useful
soon.