Eliminate leak of display lists from plug-ins
Garbage collection of display lists created by plug-ins was broken.

When wings_dl scans the #dlo{} record for display lists, it by design
does not scan arbitrary tuples to avoid finding false positives in
gb_trees or arrays.

Since #dlo.plugins contains a list of tuples, display lists are not
scanned. We'll need to update wings_dl to look into the tuples. We
also need to update the plug-ins so that they stop wrapping the display
lists in tuples (the wrapper tuple serves no useful purpose).