Raise the minimum requirements for Wings
We want to start using new features of OpenGL. Ultimately we would
want to raise the version to 4.x. For the moment we will raise the
version to 2.1. The 2.1 version is the last version that don't
deprecate immediate mode commands and the fixed functionality
pipeline that Wings still use. 2.1 is available on all modern
computers.

We also want to simplify plug-ins that use frame buffer objects
(currently the Ambient Occlusion and AutoUV plug-ins), so we will also
require that the GL_EXT_framebuffer_object is supported.