wings_dl: Replace extra/3 with draw/3
wings_dl:extra/3 was hastily introduced as a helper to eliminate a
resource leak in wings_vec (7d16c2b13a8).

After experimenting with rewriting more code to use extra/3, it turns
that in practice we always want to call wings_dl:call/1 immediately
on the return value of extra/3. Therefore, it makes more sense to
have wings_dl:draw/3 that will draw the graphic thingy after creating
it.

It also turns out that garbage collecting the "extra" map in the same
way as all other drawable terms was a mistake, because if the users
moves the camera, there would be no garbage collection and many
instances of the groundplane could be kept in the "extra"
map. Therefore, we must deallocate VBOs as soon as a key is deleted
from the "extra" map.