Refactor hotkey handling
Now that we have two sorts of menus -- native WX menus and our
popup menus -- we will need to format hotkeys differently
depending on in which kind of menu they will be used.

Therefore, change wings_hotkey:matching/1 to return unformatted
hotkeys. Add a wings_hotkey:format_hotkey(Key, Style) to format
hotkeys in the 'wx' style or in the 'pretty' style.

The Style argument only makes a different on OS X. On other platforms,
the hotkeys are always formatted in the same way.