Fix Ctrl/Command confusion on Macs
WX swaps the Ctrl and Command keys in keyboard and mouse events
(probably in an attempt to help writing code that works on both
Windows and OS X). We don't want that swapping because Wings does
a similar swapping itself when appropriate.

Therefore, swap back the keys in mouse and keyboard events when
translating from WX events to "SDL" events.