Fix crash in AutoUV for objects with virtual mirror
wings_draw_setup did not initialize #vab.data when a #vab{}
record for an empty binary was created. That is usually not
a problem because #vab.data is only used when picking and
usually we never attempt to pick an empty object. Except in
AutoUV.

The bug was introduced in aa3007d.