Eliminate use of wings_os_type in process dictionary
As a very slight optimization, the return from os:type/0 is
cached in 'wings_os_type' in the process dictionary.

Modern version of the run-time system creates a constant out of
the return value from os:type/0, so calling os:type/0 will not
create any garbage.

Any problem with using the process dictionary is that we might
want to use more processes in the future, and any variables
used in the process dictionary would have to be copied to each
process that uses them.

Therefore, get 'wings_os_type' and call os:type/0 directly.