Teach Tweak mode to play well with all camera modes
Use wings_msg:free_rmb_modifier/0 to find an unused modifier for
the current camera mode. Similarly, use the appropriate formatting
commands in wings_msg to ensure that the status line shows the
correct modifiers and mouse keys.

While we are it, do some minor refactoring and clean-up of
indentation.