Fixed the strange black line starting from mini axis
There was a strange black line being drawn in the background starting from
mini axis. Thanks to Asticles for report it. [Micheus]

The problem was found in a non proper elements calculation. The data
in Data variable is composed first by two(2) pairs (color,vertex) for every one
of the three(3) axis plus four(4) pairs (color,vertex) for every one of the
three(3) axis to build the arrow, or four pair for every one of the two axis
if the camera is looking Along of one axis.