Hotkeys option is back to working for Tweak options
Since v2.0 it was not possible to set hotkeys for Tweak option. Thanks Fonte Boa

OBS:
1) The change to wings_menu was needed to ensure the hotkey string is placed in
the right menu option as well as with the correct value;
2) Due to a change in the way to bind hotkeys (checkinf for Unicode value)
there is no way to use keys like F1-F12. Most of then used to be set for tweak
purposes by default. The previous one already set and present in the preferences
file are still working, but new ones or a "re-set" cannot be done anymore. I think
this need to be reviewed.