Fixed the issue when user try to hotkey a RMB command
OBS: The changes implemented to hotkey module in the new v.2.0.2 was prepared
to manage the hotkey (into ets) in a way different from that used to check it.
For RMB commands the bindkey/2 was returning a tuple including the selection
mode (like {bindkey, Mode, Key}), but the routine that process the hotkey was
looking into ets by a tuple without it (like {bindkey, Key}). As a workaround
I removed the Mode information and now it's working as it used to.

NOTE:
It was not possible set a hotkey for commands assigned to RMB. Thanks Fonte Boa