Refactor and stabilize cut-connect algorithm
Have a fallback when choosing the cut-plane, if average normal is bad.

Change algorithm to walk between VS to VE and make cuts and connects
where needed. The previous algorithm could make additional cuts and
could fail, also didn't reuse exising vertices in the plane if
they existed.

It still fails sometimes to find a way, it can probably be improved.