Made sure user cannot assign multiple hotkeys to the same command
It was added a dialog asking the user about to replace a hotkey when a command
already have a hotkey assigned to it. If the user accept the old hotkey will
be replaced for the new one.

That can avoid confusion when user intend to replace a hotkey and he/she was
in truth assigning a new one to the command. As we can show only one hotkey
text in the menu options it doesn't make sense we allow multiple hotkeys
assined to only one command.

The code can manage any multiple definitions already existent without crash.