Fixed the application check in render plugins.
It was fixed the render plugins code in order to look for the executable in
the provided path and check if it's valid. That is used to - in accord with
'dialogs' field in Plugins Preferences dialog - show or hide the menu options
under Render menu as well as the material settings tab in the Edit Material
dialog. That works when 'dialogs' value is 'auto'.

Due some variables be used locally and the current implementation (wx) call
the menues routine from other process, it was needed to change the old approach
to store them. As a easier option I decided to create an ets list for the
plugin module.

Made some code cleanup and reviewed the .lang files.