Implement save/load window layout
Save (and load) both contained and not contained windows.
Store them in different lists so we can import old settings from old pref
files.

Works ok, but some window size does not count surrounding borders so
windows increase each load :-/
Window borders change size depending on OS's and versions so we can
not hardcode them either.

From Micheus

That is a suggestion of fix to the layout issue.

By tracking the code it was not possible to find a way the splitters be
arranged in the saved location. The way the windows are created doesn't
allow us to do that. So, as an alternative, after all the windows has been
created and arranged in the proper layout we call a new function that set
the splitter position correctly.

It was also include a few lines to save the window size that is of interest
of us to be restored next time Wings3d is started. It's preserve its old
location is also desired, but not included in this branch.

Make all windows use the user selected font

Add wings_frame:set_title

Enable Develop menu crashes otherwise.

Fix badarith when with becomes zero in wings_palette