Fixed language files issue
The problem was only in the text for the "How To Define Hotkeys" and
it was related to changes made to the process sequence and its description
in the English file that should has been updated in those languages file that
had already a translation for the old process sequence.

NOTE: Fixed a crash caused by bad formated text for "How To Define Hotkeys"
instructions in language files and updated their description to the new
process sequence. Thanks to tkbd.