Added processing feedback to gen_char_hrl script
When a language file is updated and something is wrong/missed it's hard to
find which file is causing the problem. We can only know in which line the
problem is, but not the file name. This updates try to make this task easier
allowing us to know which was the last file processed when the crash happened.