Simplify macosx/make_dmg
We can use the -srcfolder option for hdiutil to copy the contents
onto the DMG while creating it. That will eliminate the need for
mounting the DMG to copy the stuff onto it.

Also change the compression type from ZLIB to BZIP2 for a slight
decrease in the size of the DMG.