Added support to Environment lights
Added new POVRay support to ambient light in Wings3d. Now it's possible to
use the environment lights resource available in POVRay v3.7.
It was added options to create Gradient, HDRI, SkyDome and SkyBox. For this
last one it's necessary have a light source in a scene otherwise we get a dark
scene. For the others that is not necessary.

This resource was added using as reference an article by Friedrich A. Lohmüller:
http://www.f-lohmueller.de/pov_tut/backgrnd/p_sky10.htm