It was get rid of extra 'default' material replication
As the 'default' must not be changed there is no need for create a copy
of it using a sequence number each time we open a project in a Wings3d
installation which uses a different theme. It becomes annoying when you
are sharing projects since we get a list of material like default, default1,
default2 and so on.