Fix sculpt mode and disable toolbar icons
Scuplt had the following issues (see github #169)

After enter in Sculpt mode the button on the top bar doesn't response correctly:

The selection mode buttons can be all selected or deselected at
once. After switch application and comeback to Wings3d the buttons are
updated and only the Face one is active (but it still can be
deactivated);

The axis and grid buttons doesn't update the geometry screen instantly
- it's necessary to move the camera or do any other action on the
screen to get it updated;

View or Shader button doesn't have any effect over the scene;

Save, Load, Undo and Do don't have any action.

In truth, even in normal operation the selection buttons can be all
deactivated. I believe at least the last one must be selected (down).

The combo key to activate the context menu also needs to be changed to
be consistent with the change made to Tweak - it need to be activated
by ALT+RMB instead of the old CTRL+LMB.