Added preview resource to Camera dialog
It was added preview capability to the Camera dialog. The Image Plane preview
is shown on the Geometry window as a blue border. It also can be activated by
enabling it using the proper option in View->Show menu.
It was also set some new ranges in order to avoid bad behavior. For instance
FOV >= 180° was causing Wings3D freeze and increase its memory consumption.

Note: Added a Camera settings preview. Thanks tkbd for the suggestion.