Fix quit from dock menu and release scipts
We need to veto the quit otherwise a SIGSTOP is sent to erlang process,
and we hang before we have time to save model, if quit is called from
dock menu on mac.

Also made changes needed for erlang/otp 19
Erlang now refuses to start if not BINDIR is an absolute path.
Tried to fix that.
And child_setup changed named so fixed the release script