Handle gl redraw inside callback on mac
This is used to avoid an eternal loop which happens when you start a
modal dialog (error/info box) at the same time the redraw request
appears. By using a callback to erlang these seems to not happen.

This might be an wx issue that wx handle other commands when
inside a callback.