Avoid unnecessary motion events from warp
By avoiding the warp when we already are in the position we are
warping to.

This was seen as an eternal loop on mac, the only events we got was
the motion event from warp, which became a loop.
warp(X,Y) -> motion(X,Y) -> warp(X,Y) -> motion(X,Y) -> ...
where X and Y stayed constant.