Simplify the parse transform for internationalization
It is no longer necessary for the parse transform to insert the
function name into the key tuple, because it can be handled by
the ?FUNCTION_NAME macro introduced in OTP 19. We will continue
to use the parse transform to check for errors (such as duplicated
keys), but it can be simplified.