Add a image command to converts an image to a normal-map
The original code was adapted to compute a grayscale or coloured image and
taking care about tillable images avoiding seams artefacts. It was preserved
the default scale(bumpiness) value of four.

The context menu for images on Outiline window will showup the menu item
with two options: RMB-the default one and LMB-that prompt user for Bumpiness,
Invert X and Invert Y parameters. These two last paramters allow users to
invert the bump effect.