Do not block wx thread for too long
Using wx:batch() blocks the wx thread waiting for more input from
erlang, this makes it faster sending commands to the thread.
On the other hand it blocks driver from doing concurrent work such
as forwarding events up to erlang.

So do not use wx:batch(..) on toplevel i.e. when all calculations are done,
use it only when redrawing the geom windows instead.
(That is already implemented).