Workaround to avoid the Geometry Graph window closes unexpectedly
It's hard to find in which situation the crash happens, so it was implemented a
workaround trying to avoid the hard crash of Geometry Graph window as well as it
was added a message to be send to Log windows in order to try figure out what is
the situation in which it happens.

NOTE: The Geometry Graph window closes unexpectedly in some folder/item
operations. Avoid the crash and its cause should now
be appended to Log window. Thanks to tkbd [Micheus]