Optimize wings_draw:split/3
The remove_stale_ns/2 function called by wings_draw:split/3 is
quite expensive because it rebuilds the entire array. Most of the
time, all the work is wasted because there are no stale normals
to remove. The exception is the Intrude command.

remove_stale_ns/2 is needed for the Intrude command because
make_static_edges/2 is given a list a list of the dynamic faces and
assumes that any face that is not dynamic is static. If the stale
normals have not been removed, the edges for the deleted faces will be
shown when executing the Intrude command.

Fix this by letting split_faces/4 return the face tab for the static
faces. Given the static face tab, remove_stale_ns/2 is no longer
necessary and make_static_edges/2 can be drastically simplified.