Fixed load from root dir on Windows
It has been reported situations when loading wings files we get "Read failed:
no such file or directory" and when loading images the "Failed to load "(The
tga image file's Path)": unknown POSIX error".

This was caused by that wxFileDialog:getDirectory does not add \ after
drive letter when on root dir. filename:join interprets this to be
drive relative path and not an absolute path.