Fixed the eventual fail when load Wings3D or image files
It has been reported situations when loading wings files we get "Read failed:
no such file or directory" and when loading images the "Failed to load "(The
tga image file's Path)": unknown POSIX error". The last one occurs for any
image file, but for tga files it also can be caused by user trying to load a
RLE compressed file that seems to be a wxWidgets bug.

The fix allow me to read files I was not able to load before - mainly if them
where stored in a stick memory drive.