wings_sel: Add new API functions
In the future, we want to have one process per #we{} to be
able to handle huge models with many objects.

As a first step towards that goal, we will need new API functions
that helps us avoid passing #we{} records and selections across
process boundaries. We will also need APIs that eliminate direct
access to 'sel' and 'shapes' in #st{}.

Here is an example of code that uses the existing APIs:

  {St1,Sel} = wings_sel:fold(
    fun(Items0, #we{id=Id}=We, Acc) ->
       .
       .
       .
       Items = ...,
       [{Id,Items}|Acc]
    end, [], St0),
  St = wings_sel:set(SelMode, Sel, St1)

This code cannot be split across multiple processes without
copying terms between processes.

Here is the code rewritten to use the new APIs:

  St = wings_sel:update_sel(
    fun(Items0, We) ->
       .
       .
       .
       Items = ...,
       Items
    end, SelMode, St0)

Besides being simpler, the code will be possible to parallelize
in a future commit.

Here are all the new API functions added in this commit:

  map_update_sel(F, Mode, St0) -> St.
  map_update_sel(F, St0) -> St.

  update_sel(F, Mode, St0) -> St.
  update_sel(F, St0) -> St.

  new_sel(F, Mode, St0) -> St.

The wings_sel:fold/3 and wings_sel:mapfold/3 in the existing
API should be avoided if possible. (Currently, all uses of them
are not possible to eliminate. There will be still more API
functions added in a future commit to help replacing them.)

While we are it, add specs to all exported functions. Also move
local functions to the end of the file, to make it easier to find
the API functions.