wings.hrl: Add 'temp' field to #we{}
Some complicated commands such as edge extrusion needs to be run
in several passes with extra information passed between the passes.
We don't want the extra information to be copied to the main processes
only to be copied back to the #we{} process in the next pass.

We could add more API functions in wings_sel and wings_drag that
allow the extra information to be passed. But the number of new
API functions needed to be added could soon become unwildy. Also,
most commands don't need this extra functionality, so having all
high-order functions in wings_sel and wings_drag pass along an
extra term would be inconvenient.

Instead, I choose to add an extra field in #we{} that can be used
for passing temporary information between passes.