wings.hrl: Move out types for faces, edges, vertices
It is no longer recommended to define types globally in header files.
The recommended way is to define them in an appropritate module and
export them.

Start by moving out the types for face, edge, and vertex numbers
from wings.hrl to wings_face, wings_edge, and wings_vertex,
respectively.