Added a window interface for Snap Image command.
It was created a new window and migrated the old wpc_snap code to fit on it.
The suggestion was made by dgud in this post:
http://www.wings3d.com/forum/showthread.php?tid=844&pid=11082#pid11082

OBS: maybe new will need to add text controls in order to user set manually
the exact values for Opacity and Rotation. By now thre is only sliders.