Fixed PS/EPS importer which was not importing most the user's file
Besides the crash that was reported previously, the plugin was not managing
correctly the file data with multiples objects definition. Some creators use
to fill the boundary data and others not. The code used from wpc_tt.erl module
needs that information in the right way, so I tried to prepare/repare the data
before call that module. We try avoid to load the boundary data since it results
in a big box object on screen. It was tried to avoid Wings3D crashes, se there
are some code to enable that and I also included more feedbacks to the user.

NOTE: Fixed PS/EPS importer which was not importing files properly.
It should import Adobe Illustrator, LibreOffice and Inkscape files correctly.
Scribus file may not to have some objects loaded. Thanks to tkbd for all support.