Bringing back the option Show/Hide background image to AutoUV Editor
It was added option to the main menu View|Show in to enable the user to toggle
between Show/Hide texture image in the AutoUV Editor. The menu item is added
always a AutoUV Editor window gets focused and removed when it loses the focus.

NOTE: The AutoUV Editor option to show/hide the background image is back.