Fixed autosave/save actions when storing files using a network address
Autosave was not working when the preference_dir was set to a network address
(pipe names). It was notice the save action would also cause error if the user
was selecting a folder in another PC in his network.
It was found that filename:join/2 was replacing the first "//" for "/" in a
network address like "//NEO/user/public"

NOTE: Fixed autosave error when storing files using a network address. Thanks
to broozar.