Fixed the missing call to post hook in auto rotate mode
It was noticed the redraw function in auto rotate code (in wings_view module)
was not calling an user defined draw hook (via register_postdraw_hook).
When searching for render/1 calls I found that wings_tweak module has missed
it too.
As a simple fix, the function call_post_hook/1 in wings.erl was just replicated
to those modules (which I think it's not the right action).