Teach Collapse command to remove holes for deleted faces
https://github.com/dgud/wings/issues/253

NOTE: When collapsing faces, edges, or vertices around holes, the
hole faces could be deleted, but still be marked as holes. That
would lead to a crash when attempting to save the file or when
using the Combine command. The Collapse command now makes sure
that holes for any deleted faces are removed.