Fixed missed opengl window management for customised windows in plugins.
Theached the window management module (wings_wm) to understand that a custom
OpenGL window (usually named 'geom') created by a plugin needs to respond
to mouse over events just like a regular geom/{geom,<num>}/autouv window.

That ensures the user doesn't need to click on the window's work space to get
it focused as well as it can respond to hotkeys too - since it's now focused.

Anything esle is changed - it's not necessary to change the ?IS_GEOM macro or
is_geom/1 function.