Fixed a situation in which holes can contain garbage
It was found two points where the #we{}.holes would keep garbage content.
We already had a couple of bug reports related to renumber process and holes
or hidden faces. I hoppe this will avoid any other hidden condition for crash.

NOTE: Fixed the cause of Wings3D crash after a combination of hole and
Cleanup command. Thanks to Hank.