Fixed an issue which was rollbacking the #st{} in sculpt mode
The problem stated after the UI migration. Before the main menu was disabled
in Sculpt mode and that couldn't happens.
It now can deal with that as well as the synchronization between geometry
windows is better. Originally, the sculpt states was storing a #st{} structure
three times. After the changes, it can now keep only two of them.
I avoided to make a drastic change in the code, so maybe in the future we could
evaluate a way to store only one #st{} structre ('stw' - the internal work one)
as well as remove any event that became obsolete now.
I tested the save/restore windows; the File menu options with the plugin active
and no problem was verified.

NOTE: Fixed the problem that was causing geometry state rollback after
save a project in Sculpt mode. Thanks to tkbd.