Fixed crash in Select By Short Edge command
The select by short edge was causing Wings3D to crash dua a wrong function
bind. There is no need to restrict the command to unselected objects.
In fact, the code was already ready to deal with partial selections and
look for short edges only in selections.

NOTE: Fixed the cause of Select By Short Edge be crashing when there was
a active selection. Thanks to rv3.