Fixed the docking issue when the user has multiple monitors
Docking was not working in a multiple monitors layout. The MouseState function
didn't take them in account properly. For example, for a secundary monitor at
left the X coordinate less than 0 has a value like 4294997294 equivalent to -1
and 4294965376 equivantet to -1920.
The function getMousePosition() returns the proper values fixing the issue.

NOTE: Fixed the window docking feature for multiple monitors systems. Thanks to suzuki.